% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-contrast.R
\name{as.matrix.dkge_contrasts}
\alias{as.matrix.dkge_contrasts}
\title{Extract contrast values as matrix}
\usage{
\method{as.matrix}{dkge_contrasts}(x, contrast = 1, ...)
}
\arguments{
\item{x}{A dkge_contrasts object}

\item{contrast}{Name or index of contrast to extract}

\item{...}{Additional arguments (not used)}
}
\value{
S×P matrix of contrast values
}
\description{
Extract contrast values as matrix
}
