% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-package.R
\docType{package}
\name{dkge-package}
\alias{dkge-package}
\title{dkge: Design-Kernel Group Embedding}
\description{
Core routines for design-aware group embedding, cross-validated contrasts, and transport utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/dkge/}
  \item \url{https://github.com/bbuchsbaum/dkge}
  \item Report bugs at \url{https://github.com/bbuchsbaum/dkge/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
