% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-neuroim2.R
\name{dkge_cluster_ts}
\alias{dkge_cluster_ts}
\title{Aggregate voxel time series into cluster means}
\usage{
dkge_cluster_ts(bv, labels, ids = NULL, chunker = NULL)
}
\arguments{
\item{bv}{A \code{neuroim2::NeuroVec} object (4D time-series data, TxXxYxZ).}

\item{labels}{A \code{neuroim2::NeuroVol} with integer cluster identifiers
(0 indicates background).}

\item{ids}{Optional subset of cluster IDs to retain.}

\item{chunker}{Optional function returning list(mat = TxV_block, vox_idx).}
}
\value{
Matrix of dimension TxP where P is the number of clusters retained.
}
\description{
Aggregate voxel time series into cluster means
}
\examples{
\donttest{
if (requireNamespace("neuroim2", quietly = TRUE)) {
  labels <- neuroim2::read_vol(system.file("extdata", "global_mask2.nii.gz", package = "neuroim2"))
  vols <- lapply(1:4, function(i) labels * i)
  bv <- neuroim2::vec_from_vols(vols)
  ts <- dkge_cluster_ts(bv, labels)
  dim(ts)
}
}
}
