% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-alignment-api.R
\name{dkge_paint_medoid_map}
\alias{dkge_paint_medoid_map}
\title{Paint medoid values back to a label volume (deprecated)}
\usage{
dkge_paint_medoid_map(values, labels, out_file = NULL)
}
\arguments{
\item{values}{Vector or matrix of values defined on the reference labels.}

\item{labels}{Labels describing the identified target support.}

\item{out_file}{Optional output path.}
}
\description{
Paint medoid values back to a label volume (deprecated)
}
