% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-project.R
\name{dkge_project_blocks}
\alias{dkge_project_blocks}
\title{Project new blocks into DKGE score space}
\usage{
dkge_project_blocks(fit, B_list, Omega_list = NULL, w = NULL)
}
\arguments{
\item{fit}{A \code{dkge} object.}

\item{B_list}{List of subject beta matrices.}

\item{Omega_list}{Optional list of spatial weights aligned with \code{B_list}.}

\item{w}{Optional vector of subject weights for the new data.}
}
\value{
Matrix of projected scores (qxrank).
}
\description{
Project new blocks into DKGE score space
}
\keyword{internal}
