% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-project.R
\name{dkge_project_cluster}
\alias{dkge_project_cluster}
\title{Project a new cluster/voxel vector onto DKGE components}
\usage{
dkge_project_cluster(fit, b, omega = 1, w = 1)
}
\arguments{
\item{fit}{A \code{dkge} object.}

\item{b}{Numeric vector of length q (effects).}

\item{omega}{Optional scalar or matrix weight.}

\item{w}{Optional scalar subject weight.}
}
\value{
Numeric vector of length \code{rank} (component scores).
}
\description{
Project a new cluster/voxel vector onto DKGE components
}
\keyword{internal}
