% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-services.R
\name{dkge_run_service}
\alias{dkge_run_service}
\alias{dkge_run_service.dkge_contrast_service}
\alias{dkge_run_service.dkge_inference_service}
\alias{dkge_run_service.dkge_transport_service}
\alias{dkge_run_service.default}
\title{Run a dkge service object}
\usage{
dkge_run_service(service, ...)

\method{dkge_run_service}{dkge_contrast_service}(service, ...)

\method{dkge_run_service}{dkge_inference_service}(service, ...)

\method{dkge_run_service}{dkge_transport_service}(service, ...)

\method{dkge_run_service}{default}(service, ...)
}
\arguments{
\item{service}{A service object created by \code{dkge_contrast_service()},
\code{dkge_inference_service()}, or \code{dkge_transport_service()}}

\item{...}{Additional arguments passed to the typed runner}
}
\value{
Result from the typed service runner
}
\description{
Generic dispatcher for executing pre-configured service objects. Dispatches
to the appropriate typed runner based on the service class.
}
