% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-kernel.R
\name{helmert_contrasts}
\alias{helmert_contrasts}
\title{Helmert contrasts for a set of factors}
\usage{
helmert_contrasts(Ls)
}
\arguments{
\item{Ls}{Numeric vector of factor levels (named or unnamed).}
}
\value{
Named list of orthonormal Helmert contrast matrices.
}
\description{
Helmert contrasts for a set of factors
}
