% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-kfold.R
\name{print.dkge_folds}
\alias{print.dkge_folds}
\title{Print method for dkge_folds}
\usage{
\method{print}{dkge_folds}(x, ...)
}
\arguments{
\item{x}{A dkge_folds object}

\item{...}{Additional arguments (unused)}
}
\value{
`x`, invisibly.
}
\description{
Print method for dkge_folds
}
