% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-kernel.R
\name{sum_contrasts}
\alias{sum_contrasts}
\title{Sum-to-zero contrasts for a set of factors}
\usage{
sum_contrasts(Ls)
}
\arguments{
\item{Ls}{Numeric vector of factor levels (named or unnamed).}
}
\value{
Named list of Lx(L-1) contrast matrices.
}
\description{
Sum-to-zero contrasts for a set of factors
}
