% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-render-core.R
\name{dkge_anchor_to_voxel_apply}
\alias{dkge_anchor_to_voxel_apply}
\title{Decode anchor values to voxel space}
\usage{
dkge_anchor_to_voxel_apply(decoder, anchor_values)
}
\arguments{
\item{decoder}{Object returned by [dkge_anchor_to_voxel_fit()].}

\item{anchor_values}{Numeric vector of length `decoder$n_anchors`.}
}
\value{
Numeric vector of voxel values.
}
\description{
Decode anchor values to voxel space
}
