% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-neuroim2.R
\name{dkge_cluster_betas}
\alias{dkge_cluster_betas}
\title{Compute cluster-level betas from neuroim2 objects}
\usage{
dkge_cluster_betas(bv, x_mat, labels)
}
\arguments{
\item{bv}{A `neuroim2::NeuroVec` object containing 4D time-series data.}

\item{x_mat}{Subject design matrix (Txq).}

\item{labels}{A `neuroim2::NeuroVol` object with cluster label assignments.}
}
\value{
Matrix of GLM betas (qxP) where P is the number of clusters.
}
\description{
Compute cluster-level betas from neuroim2 objects
}
\examples{
\donttest{
if (requireNamespace("neuroim2", quietly = TRUE)) {
  labels <- neuroim2::read_vol(system.file("extdata", "global_mask2.nii.gz", package = "neuroim2"))
  vols <- lapply(1:5, function(i) labels * i)
  bv <- neuroim2::vec_from_vols(vols)
  x_mat <- cbind(intercept = 1, trend = seq_len(length(vols)))
  betas <- dkge_cluster_betas(bv, x_mat, labels)
  dim(betas)
}
}
}
