% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-predict.R
\name{dkge_freeze}
\alias{dkge_freeze}
\title{Freeze a DKGE fit into a compact model for prediction}
\usage{
dkge_freeze(fit)
}
\arguments{
\item{fit}{a dkge or dkge_stream object}
}
\value{
list with U, K, R and class 'dkge_model'
}
\description{
Freeze a DKGE fit into a compact model for prediction
}
