% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-services.R
\name{dkge_inference_service}
\alias{dkge_inference_service}
\title{Construct an inference service}
\usage{
dkge_inference_service(spec = NULL, ...)
}
\arguments{
\item{spec}{Inference specification (list or `dkge_inference_spec`).}

\item{...}{Additional parameters merged into the specification.}
}
\value{
Object of class `dkge_inference_service`.
}
\description{
Construct an inference service
}
\examples{
inference_srv <- dkge_inference_service(dkge_inference_spec(B = 1000))
}
