% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-plot.R
\name{dkge_plot_effect_loadings}
\alias{dkge_plot_effect_loadings}
\title{Effect-space loadings heatmap (K \%*\% U)}
\usage{
dkge_plot_effect_loadings(fit, comps = NULL, zscore = FALSE)
}
\arguments{
\item{fit}{Fitted `dkge` object.}

\item{comps}{Components to include (defaults to first min(rank,6)).}

\item{zscore}{Logical; z-score loadings within each effect.}
}
\value{
A ggplot object.
}
\description{
Effect-space loadings heatmap (K \%*\% U)
}
