% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-plot.R
\name{dkge_plot_info_anchor}
\alias{dkge_plot_info_anchor}
\title{Anchor-level information plots}
\usage{
dkge_plot_info_anchor(info_haufe = NULL, info_loco = NULL, top = 20)
}
\arguments{
\item{info_haufe}{Result from `dkge_info_map_haufe()`.}

\item{info_loco}{Result from `dkge_info_map_loco()`.}

\item{top}{Number of anchors to annotate.}
}
\value{
List of ggplot objects.
}
\description{
Anchor-level information plots
}
