% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-plot.R
\name{dkge_plot_scree}
\alias{dkge_plot_scree}
\title{DKGE scree plot with cumulative curve}
\usage{
dkge_plot_scree(fit, one_se_pick = NULL)
}
\arguments{
\item{fit}{Fitted `dkge` object.}

\item{one_se_pick}{Optional integer component chosen by one-SE rule.}
}
\value{
A ggplot object.
}
\description{
DKGE scree plot with cumulative curve
}
