% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-predict.R
\name{dkge_predict_loadings}
\alias{dkge_predict_loadings}
\title{Predict DKGE loadings for new subjects (out-of-sample)}
\usage{
dkge_predict_loadings(object, B_list)
}
\arguments{
\item{object}{dkge | dkge_stream | dkge_model}

\item{B_list}{list of qxP_s beta matrices for new subjects}
}
\value{
list of P_sxr loadings (A_s) for each subject
}
\description{
Predict DKGE loadings for new subjects (out-of-sample)
}
