% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-predict.R
\name{dkge_predict_stream}
\alias{dkge_predict_stream}
\title{Streaming prediction for new subjects via a loader}
\usage{
dkge_predict_stream(object, loader, contrasts)
}
\arguments{
\item{object}{dkge | dkge_stream | dkge_model}

\item{loader}{object with n(), B(s) methods (and optional X(s))}

\item{contrasts}{list or matrix as in dkge_predict()}
}
\value{
list(values=list per subject, A_list=list of loadings)
}
\description{
Streaming prediction for new subjects via a loader
}
