% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-project.R
\name{dkge_project_clusters}
\alias{dkge_project_clusters}
\title{Project multiple cluster/voxel vectors}
\usage{
dkge_project_clusters(fit, B, omega_vec = NULL, w = 1)
}
\arguments{
\item{fit}{A `dkge` object.}

\item{B}{qxP matrix of cluster betas.}

\item{omega_vec}{Optional vector of per-cluster weights.}

\item{w}{Optional subject weight.}
}
\value{
Pxrank matrix of projected coordinates.
}
\description{
Project multiple cluster/voxel vectors
}
