% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-services.R
\name{dkge_transport_service}
\alias{dkge_transport_service}
\title{Construct a transport service}
\usage{
dkge_transport_service(spec = NULL, ...)
}
\arguments{
\item{spec}{Transport specification (list or `dkge_transport_spec`).}

\item{...}{Additional key-value pairs merged into the specification.}
}
\value{
Object of class `dkge_transport_service`.
}
\description{
Construct a transport service
}
\examples{
transport_srv <- dkge_transport_service(dkge_transport_spec(centroids = list(matrix(0, 2, 3))))
}
