% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-weights.R
\name{dkge_weights_auto}
\alias{dkge_weights_auto}
\title{Default DKGE weight specification}
\usage{
dkge_weights_auto()
}
\value{
`dkge_weights()` object with k-energy precision combination.
}
\description{
Default DKGE weight specification
}
