% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_sqdist_cpp}
\alias{pairwise_sqdist_cpp}
\title{Pairwise squared distances between two point sets (rows)}
\usage{
pairwise_sqdist_cpp(A, B)
}
\arguments{
\item{A}{nxd}

\item{B}{mxd}
}
\value{
nxm matrix of ||a_i - b_j||^2
}
\description{
Pairwise squared distances between two point sets (rows)
}
