% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-contrast.R
\name{print.dkge_contrasts}
\alias{print.dkge_contrasts}
\title{Print method for dkge_contrasts}
\usage{
\method{print}{dkge_contrasts}(x, ...)
}
\arguments{
\item{x}{A dkge_contrasts object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dkge_contrasts
}
