% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkge-inference.R
\name{print.dkge_inference}
\alias{print.dkge_inference}
\title{Print method for dkge_inference}
\usage{
\method{print}{dkge_inference}(x, ...)
}
\arguments{
\item{x}{A dkge_inference object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dkge_inference
}
